+++ ChromeExtension/content-script.js
@@ -147,6 +147,14 @@ function cran_inject_other_urls() {
     a.title = "Package page on Microsoft MRAN";
     td.appendChild(a);
 
+    td.appendChild(document.createElement("br"));
+    a = document.createElement("a");
+    url = "https://rdrr.io/cran/" + pkg;
+    a.innerText = url;
+    a.href = url;
+    a.title = "Package page on rdrr.io";
+    td.appendChild(a);
+
     tr2.appendChild(td);
     table.appendChild(tr2);
 }